+++ changelog.md
@@ -9,6 +9,7 @@
 * Add support for FreeBSD
 * Drop support for Windows 32 bit
 * [Issue 287](https://github.com/jacob-carlborg/dstep/issues/287): Add support for adding public imports
+* [Issue 288](https://github.com/jacob-carlborg/dstep/issues/288): Missing `__u/int128` support
 
 ### Bugs Fixed